Edited front-end of Create Post page
I changed the background of the form to white to align with the other pages.

I also added a box shadow to the page. Debating doing this for the other pages.

Removed the return to home button and replaced it with the ReMarket Logo which is now used to go back to the homepage.

TODO:
Remove "Contact Name" and "Contact Number" fields from the form
modify create_post.php to accomodate this. Also update it to store item names.