Created Cart UI Element for Homepage
Used Javascript and HTML to create Cart UI element with a dropdown menu.

I also made ReMarket logo in index.html refresh the page.

I put the logo in index.html in its own class so that it is uniform with the UI elements.

Made the search bar smaller.

TODO:
In index.html:
Use PHP to read and display cart items in the dropdown menu
Use PHP to check if there are cart items in the database. If there are none, display "No items are in your cart!" I have a temporary placeholder for now.