Work on Home Page and Header
The Home Page is now in the process of being ported to Tailwind. It now supports dark mode, for example.
It just looks better.

I have similarly started work on the header. The shopping cart now works as a slide over. It's not even close to finsihed though.

I have also switched us from Tailwind Play to Tailwind CLI so that we can use custom colors.