Account User Menu Closes if clicked off
User Menu no longer requires pressing the Account Picture to toggle the menu.

I did this by adding click listeners to check if the user has clicked off of the menu.

Something weird is I had to give the account button an id and add another click listener for that. That is what finally fixed it.